nixos-render-docs: add Renderer base class
